requirements.txt
azure-ai-vision-imageanalysis==1.0.0b2
azure-core==1.30.2
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.7
importlib_metadata==7.2.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
